+++ src/app/photographer/packages/[packageID]/edit/page.tsx
@@ -1,4 +1,4 @@
-import EditPackage from '@/components/package-page/edit-package'
+import EditPackage from '@/components/photographer/package-page/edit-package'
 
 export default function EditPackagePage() {
   return <EditPackage />
