+++ src/app/photographer/packages/create/page.tsx
@@ -1,4 +1,4 @@
-import CreatePackage from '@/components/package-page/create-package'
+import CreatePackage from '@/components/photographer/package-page/create-package'
 
 export default function CreatePackagePage() {
   return <CreatePackage />
