+++ src/app/photographer/verify/page.tsx
@@ -1,4 +1,4 @@
-import VerifyPhotographer from '@/components/verify-photographer'
+import VerifyPhotographer from '@/components/photographer/verify-photographer'
 
 export default async function Page() {
   return <VerifyPhotographer />
