+++ src/components/photographer/package-page/edit-package-detail.tsx
