+++ src/components/photographer/package-page/package-detail.tsx
