+++ src/components/photographer/verify-photographer/index.tsx
